0.0.11. Implement showing a product with details.

+++ src/main/java/ru/otus/java/basic/webapi/dto/product/ProductInputDto.java
@@ -1,8 +1,10 @@
 package ru.otus.java.basic.webapi.dto.product;
 
+import java.math.BigDecimal;
+
 public record ProductInputDto(
         int categoryId,
         String name,
         String description,
-        double price
+        BigDecimal price
 ){}
